Fixed issue with analysing positional lengths in relaxation workchain (#290)
* Fixed issue with analysing positional lengths in relaxation workchain

Due to nan entries the `max` function failed and the relaxation workchain
did not work as expected. This is now fixed, we keep the nan, but act
on the raised `RuntimeWarning` from numpy and set force the convergence
to not be complete if this in encountered. An update to the plane wave
and k-point convergence workchain is also performed in order to enable a final
run using relaxation and the converged plane wave and k-point cutoffs.

* Fixed the positional convergence test