Compatibility with AiiDA 1.1.0
In this commit we clean up the package dependencies such that we strictly follow AiiDA versioning also on the development and testing part which makes it easier to perform development on both aiida-core and aiida-vasp.

In addition, liniting was redone to suite updated versions.

Note that we had to disable the immigrator tests since they got stuck forever. Solving this is left as an issue.