Generalization of workchains (#314)
In order to open for the use of the workchains in other codes, we here decouple the workchains and the VASP specific parameters. In the workchains we use AiiDA/workchain specific parameters, which is translated at the code specific workchain in the workchain stack to code specific parameters. This concept can also be utilized by other codes.