Add lambda on input defaults (#316)
To comply with the update https://github.com/aiidateam/aiida-core/pull/3466 we here incorporate lambda on all input nodes that are of an AiiDA data type. This is done to make it clear that we want to supply a immutable object as input. This also fixes the problem of not being able to run consequetive workchain tests, so in this PR all of the present workchain tests are finally enabled. In addition this PR fixes the previous problems of not being able to build RTD due to the need of a backend. Please consult https://github.com/aiidateam/aiida-core/pull/3466 and reference therein for additional details.