Make correct handling of STDOUT file in parsers
The problem with STDOUT is that its destination can be customised
by the CalcJob. Hence, we cannot hard code its filename.
To overcome this, I add special cases for processing STDOUT when
checking quantity and getting the actual file. The file name
STDOUT is just a marker for this special case.
The test case `test_vasp_calc` is updated to also test parsing of
 errors.