Use positive match for restart folder copy
Previously the copying of the restart folder will copy all files
including the output files such as vasprun.xml and OUTCAR. Which
can result in incorrect outcome if the remote job is kill.
The copy is now limited to CHGCAR and WAVECAR, and the remote folder
is checked for constructing the list.

The `write_additonal` method now simply checks the existence of the
files in the copy list.