Updated dependencies
Additional quantities can be parsed in `parsevasp>=2.0.0` and thus we update the
dependencies. In addition, we changed the extra tags when installing to comply with the standard, including `aiida-core`. During the update of the parsing of the total energies in `parsevasp` the default energy was changed from `energy_no_enetropy` to `energy_extrapolated`. Additional tests have been added, in addition to the posibilities of storing the total energies of each electronic step.

Furthermore, we have removed `prospector` and execute `pylint` directly, similar to what has been done in `aiida-core`.