Updated test configs
Removed Travis linter from the pre-commit config, updated the tox config to reflect the change of extra naming in previous commit. In addition, version limit to pymatgen is added, as version 2020.12.18 gave np.ndarray size changed errors. There is presently no version limit in AiiDA core, so it is added to the requirements here. Since we have duplicated code in the workchains, but would like to postpone refactoring of that, the pylint check of similar lines was elevated from the default of 4 to 25.