Update of parameter massager
When allowing selective dynamics flags in PR #408 some bugs where introduced which caused e.g. the flags to be not put correctly in the code namespace. This is fixed, a test is added and in addition we introduce the following specification: the namespaces allowed on the input parameters are restricted. This means that if the user
want sto override a parameter (send it straight to VASP) it needs to be put in the `incar` namespace in the workchains before the `VaspWorkChain`. For interacting with the calculations, any INCAR flag can be set in the `inputs.parameters`. If any parameter is detected on the input parameter root and it is not in the allowed namespace, an exception will be raised. While updating we also disconnected the massager from the workchain, which was previously supplied to get access to the exit codes. Now, we instead raise exceptions with a sensible text and then catch this in the VASP workchain and customize an exit code there that incorporates the exception message. Furthermore, we changes the behavior that was introduced in PR #408 such that `dynamics` is now an optional input port on `VaspCalculation` and is supplied there instead of in `parameters`. This made it possible to relax the need for additional namespaces being used in `parameters` for the `VaspCalculation`. This was initially motivated by the ongoing work to introduce a parameter checker for dict nodes, but since that work is not complete, we here instead follow standard AiiDA logic.