Make PotcatData -> PotcarFileData only match sha512
Now getting the PotcarFileData from PotcarData is only through
matching SHA512. Since the SHA512 is only computed from the file
content, matching a PotcarFileData with different other attributes
does not change the end result. In fixes inconsistency when upload
family containing POTCAR that does not exsit in the database but a
PotcarData does exist.