Initial refectoring of the parser routine
The core goal is to guard for exceptions, so exception during paring
a specific quantity/file does not except the entire VaspCalculation
process. Note that some, but not all file parsers already have
internal handling. However, ensuring each parser handles all exceptions
internally is not as robust as using a top level handling.