Added recording of file_parser_warnings
Previously, each file parser may return a exit_code, and the top
level parser simply inherient the exit_code from the last file
parser. This is not a transparent processce as parser may not happen
in certain order, plus that the method of different quantities may
return different exit_codes.

This is now changed as the exit_code from all file parsers are now
collected and stored in a `file_parser_warnings` field, making it
easier to understand which file parser went wrong. Note that the issue
of multiple quantities giving different exit_code within the same
file parser object is not addressed.