Added error code for not reaching ionic convergence
This allows the error handling to be passed through to higher level
workchain if need by disabling the ionic convergence handlers. In
this case, the `VaspWorkChain` will terminate with an error code
to indicate it.

The same goes for electronic convergence problems, by disabling the
handlers, electronic convergence issues may be handled by higher-level
workchains.