Added always_attach_outputs handler
This handler is only active in the last iteration given that  the
calcultion did not finish OK. The idea is that it attaches
the outputs of the last calculation regardlessly and skips other
handlers - this can be useful when the workflow is also controlled by a higher
level workchain.

By default, this handler is disabled - so only VaspWorkChain with
exit_code=0 returns the final outputs.