Remove `with_dbenv` decorator from `get_data_class`
The `get_data_class` utility was decorated with `with_dbenv`. This
function is supposed to load the database environment if it is not
already loaded. However, a bug in that function causes it to load the
database to be loaded everytime. This causes a simple call like:

     RelaxWorkChain.get_builder()

to last 10 seconds, instead of being almost instantaneous.

Now, the bug should of course be addressed in `aiida-core` but the
database really doesn't have to be loaded to load the data classes, so
it is best to remove the decorator entirely. The correct profile will be
instantiated automatically as soon as an instance is created of any of
the loaded data classes.

The `mock_vasp` CLI command is now decorated with the `with_dbenv`
decorator as it was relying on the database being loaded indirectly by
the `get_data_class` utility but this is no longer the case.