Use PotcarGroup for managing potcar families
Previously, user generated and plugin internal groups are separetely
by `type_string` used when creating and querying. This was changed
in aiida-core 1.2 and now the preferred way is to make a subclass
of the Group and assign an entrypoint under aiida.group for it.

The code is now refectored to use the latter approach, and a command
is added to migrate eixsting groups to the new implementation.