Refactor the handler framework
The final checks are splited into multiple ones so they can be disabled
separately.
The logic of ionic and electronic fixes are updated such that those
contains ionic steps with NELM breach will be treated as having
electronic convergence issues. If the calculation itself finishes OK,
then the workchain is aborted for further inspections.
The handler names are updated to be more concise and make it easier
to override them on/off.