Addes invalidates_cache to be true for some exit codes
Basically, if the return code can go away after a restart, the cache
should be invalidated. This includes problems on the cluster side, or
VASP got interrupted (walltime limit).

Calculations that have reached the end of execution should not have
the cache invalidated.
Errors related to parsing should not have the cache invalidated either,
as reruning will produce the same problem.