Use `expose_outputs` for VaspWorkChain
This fixes an accidiental issue with aiida-core 1.6.5. Previously,
the outputs are repeated explicitly, which is now replaced by
exposed_outputs. On aiida-core > 1.6.5, the nested outputs are now
handled correctly, and `BaseRestartWorkChain` uses this to check for
required outputs for the underlying calculation node.

Fix for the immigrate class

Changing _process_class (vasp.immigrate) and _next_workchain (vasp.vasp)
has the same outputs actually. Change this so the BaseRestartWorkChain
works as intended.