Avoid np.nan in convergence data
The results of pwcutoff and k-point convergence tests are stored in
AiiDA's Dict. Thus, we cannot use np.nan for representing failed
runs because nan values in Dict raise error when they are stored. To
avoid this situation, this fix add a new boolean entry in `pw_data` and `k_data`
that represents the corresponding run successed or not.