Mock code enhancement
Allow the MockRegistry to support multiple search paths.
In the maintime, there is still only one base_path, which is where
new uploads will be placed.

Remove printing AiiDA environment in mock. The mock code does not
know if it is run by AiiDA or not, just like a normal vasp.
Since its operation is independent of AiiDA, it does not make sense
to have it echo out AIIDA_PATH and profile information.