Deviating from develop to maintain a version that supports core v1
Here is the first commit that deviates from develop to offer support in the
future for core v1. Here we also add error handlers for the ACWF study which
is not presently in the develop branch. They are disabled by default to retain
backwards compatibility. Pylint has also been disabled.