Update dependencies
Here we update the `parsevasp` dependency to `~=3.2` and `aiida-core` to `~=2.3`.
Ina ddition we remove the `pre-commit` extra which followed `aiida-core` to include
a more recent `pylint` of `~=2.15` due to `wrapt` issues for Python 3.11.
Due to linting updates, also changes have been made to the linting configs and some
code quality improvements.