Devops: Add continuous deployment to CI workflow
The CI workflow is updated to trigger deployment of the package to PyPI
when a new tag is pushed to the main repository and the pre-commit and
tests have passed successfully. As a final check, the tag label is
compared with the version defined in the `aiida_vasp/__init__.py` to
make sure they match and prevent accidental incorrect tags.