Added keep_last_workdir option
This is the option to ensure that the last output of the workchain
should be kept from the remote folding being cleaned. Useful when
reusing the output remote folder for restart. All intermediate
calculation's remote will be cleaned as normal if clean_wordir is
True (default).

Also cleaned the very outdated examples in the ./examples/ folder.