Calculation reuse enhancement (#753)
* Added option to use WAVECAR restart for hybrid bands

* Update code for using restart folder

* Update the logic or keep_sp_workdir

This option is used to make sure that the output workdir is NOT
cleaned.
This is needed when reuse of WAVECAR/CHGCAR is needed between
called workchains.

NOTE: we should implement logic  to keep the last remote_folder
for VaspWorkChain but allow the intermediate works to be cleaned.

* Added keep_last_workdir option

This is the option to ensure that the last output of the workchain
should be kept from the remote folding being cleaned. Useful when
reusing the output remote folder for restart. All intermediate
calculation's remote will be cleaned as normal if clean_wordir is
True (default).

Also cleaned the very outdated examples in the ./examples/ folder.

* Fix