Added logic for fixing kpoint inconsistent problem
For hybrid band structure we need to have the scf kpoints. Previously
this must be provided explicitly or exacted from the relaxation
calculation. Since the actual calculated structure is standardised
by the path generation algorithm, the provided/extract kpoints may
not be consistent with the actual structure being calculated.
Added a new step to run a SCF calculation for extracting the kpoints.
This calculation is a bit wasteful but will ensure that the results
are current.
In the future we should replace this SCF with dryrun or some code
that can obtain the VASP SCF grid externally (if such code exists).