Implement monitors for VaspCalculation (#796)
* Implemented monitors for VaspCalculation

The monitors can check for errorous large stdout file (~ GB) which can
happen in rare cases where VASP got stuck. This will make the
daemon get stuck during retrieve or parsing.

Also check for very long LOOP time compared to the walltime limit
as well as those got stuck.

Add missing monitors.py

Added potential family validation

Minor fix for the documentation

Remove the use of Kpoints for type annotation as the user may not
have pymatgen installed.

* Added pymatgen as test and docs dependency

The building of documents requires pymatgen to be installed for
the evaluation of the code-cells.

Also used profiling to find the case of slow testing,
improments includes:

- Avoid excessive use of aiida_profile_clean and optimised some
  potcar upload routine.
- A cache is usd by the ParametersMassage to avoid necessary repeated
  loading of files.

* Fix ci - install tests optional dependences

* Fix errors when pymatgen is not installed

* Update monitor code to comply with aiida-core==2.7.1

The use of chdir and getcwd is deprecated. aiida-core no longer
change directory to the working directory before pass the transport.

* Switch to use uv for CI