Replace .options with calc.metadata.options
Directly expose options as calc.matadata.options for VaspWorkChain.
The use of an extra options node was due to early version of aiida-core
not able to serialize none_db inputs. This is no longer the case
since a long time ago.

Also refectored the VaspWorkChain code by using expose_inputs at
root level. This reduces redundant code and allow any changes to
the VaspCalculation to be forwarded to the workchains.