👌 Use `parameters_validator` to convert to lowercase
The `parameters` input of the `VaspWorkChain` has a validator (`parameters_validator`)
that among other things checks if all the keys of the provided dictionary are lowercase.
If they are not, an `InputValidationError` is raised.

Since it is possible to update the contents of the validated `node` in the validator,
here we adapt the validator to change the case of the provided keys to lowercase
on the fly _if the node is not stored_, and warn the user of any keys that were updated.