Various improvements and preparation for 5.0.0 release (#821)
* Further optimisations to the potcar related code

* Move v2.common and v2.inputset modules to top level

* Added aiida-vasp command line tool

* Minor update

* Replace .options with calc.metadata.options

Directly expose options as calc.matadata.options for VaspWorkChain.
The use of an extra options node was due to early version of aiida-core
not able to serialize none_db inputs. This is no longer the case
since a long time ago.

Also refectored the VaspWorkChain code by using expose_inputs at
root level. This reduces redundant code and allow any changes to
the VaspCalculation to be forwarded to the workchains.

* Checkpoint

* Add missing protocols

* Update code and add tests

* Add protocol for bands and convergence

* Allow standalone `options` for backward compatibility

It was quite a bit of hack to get it working. The `options` may be
used instead of the preferred calc.metadata.options for
VaspWorkChain. This is mainly for backward compatibility so old
scripts defining `options` port can still be used.

* Fix command module

* Add missing yaml file

* Added ProtocolUpdater class

* Initial implementation of updater

* Bug fix and modify launch.py to use protocols

* Update immigrant to use the remote_folder port

This allow VaspCalculation to be used directly for importing existing
calculations.

* Various improvements and bug fixes

* Update documentation

* Update the documentation

* Rename Updater to InputGenerator

* Added immigrant command import-calc

Also allow existing calculations to be used as overrides to
the protocols.

TODO: add mode to ignore existing preset and only use the calculation
provided.

* Added CLI references using sphinx-click

* Add sphinx click to the dependencies

* Fix tests

* Update documentation

* Added pymatgen input set compatibility layer

* Improve handling of errorous VASP potcar data set

Older releases of VASP potcar set may be labeled with incorrect
symbols. This includes the PBE/LDA set, PBE/LDA.52/54's original
release. Current releases of PBE/LDA.52/54, PBE/LDA.64 are not
affected. Previously, these errores may result in duplicated sybmols
in the potential family group, causing incorrect POTCAR mappings in
some cases. For example, X_GW/POTCAR may be labeld as X, and hence
replacing the acutal X/POTCAR when assigned.

This commit adds work arounds to fix these cases and also includes
a set of SHA512 data in order to "verify" a potential data set against
known sets downloaded from the VASP portal.

* Fix errors

Allow modifying nested sub-workchain's protocol by simply setting
`protocol` in the inputs.

Set {'vasp': {'protocol': 'balanced'}} will make a VaspRelaxWorkChain
run VaspWorkChain based on the 'balanced' protocol. This can be done
in both the protocol file itself as well as the overrides supplied.

* Updated recursive_merge function with special handlings

The following special operations may be applied:
  - ``$!del``: Delete this key from the result.
  - ``$!append``: Append a value to a list at this key.
  - ``$!extend``: Extend a list at this key with another list.
  - ``$!replace``: Replace the value at this key entirely.

* Fix

* Fix converge updater

* Fix for the conv_settings

* Add docstring

* Replace verdi data with aiida-vasp as top level commands

* Added script for fixing inconsisent POTCAR symbol mappings

* Update terminology

PBE.52.nohash refers to the original PBE.52 release without hash.
They are not  NOT the POTCARs from potpaw.PBE.52.tar.gz.

* CMD fix

* Fix the commands

* Add tests for commandline interface

* Update tests

* Update

* Bump version

* Update

* Added tests for the protocol mixins

* Simplify and fix tests

* Fix dict merging bug and improve launch

- added `--alias`
- added `--yes`
- Allow both structure and from-vasp-folder to be specified

* Fix and added test for using --match-existing

* Add tests

* Update

* Add remarks about setting positions_dof

* Append label for relax in the band workchain

* Fix incorrect use of outputs.get

* Rearrange the documentation

* Rearrange and update test

* Added test_generator.py

* Documentation checkpoint

* Reorganize the documentation structure

* Reorganize documentations

* Documentation update

* Typo fix

* Update documentation

* Hot fix

* Hot fix

* Hot fix for monitor

It appears the monitor is executed even if the job is not started
yet, e.g. queued, the monitor should take this into consideration.

* Slight update of the convergence update logic

* Added documentation for setting magmom and ldau

* Only set magmom_mapping if explicitly supplied

* Update protocol builder generator logic

* Update README.md

* Add missing IBZKPT files

* Fix tests

* Add better skip condition in cli test

* Code update based on pre-commit